api_dating_and_matches/requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
channels==3.0.4
channels-redis==3.4.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==40.0.2
daphne==3.0.2
Django==4.0.4
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
msgpack==1.0.5
Pillow==9.1.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.1.1
pytz==2023.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0

